feat: Add text size preference setting
This commit adds the ability for users to change the text size of their
chat messages.

- Added a new enum for text sizes. - Created a new request to validate
the text size. - Added a new component for the text size tabs. - Added a
new composable for handling the text size. - Updated the UserPreference
model. - Added a new route for updating the text size preference. -
Updated the profile controller.